requirements.txt
flask
matplotlib
numpy
pandas
scikit-learn
gunicorn
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
dotenv